pyproject.toml
[tool.poetry]
name = "finch"
version = "0.1.0"
description = ""
authors = ["Stevan Vlajic <91951603+stevan06v@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai-whisper = "^20250625"
setuptools-rust = "^1.12.0"
yt-dlp = "^2025.10.14"
pydantic = "^2.12.2"
pydub = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
